+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.5.0] - 2025-03-07
+
+### Changed
+
+- [dashboard] Using 1.6.0 with updated UI (About page)
+- [reverse-proxy] Using 1.11 with custom kiwix external link block page
+
 ## [2.4.2] - 2025-03-07
 
 ### Fixed